+++ commonjs/src/sdk/appcelerator.js
@@ -1,0 +1,15 @@
+/**
+ * @preserve
+ * Titanium Cloud Module
+ *
+ * This module is used across three distinct platforms (iOS, Android, and Mobile Web), each with their own architectural
+ * demands upon it.
+ *
+ * Appcelerator Titanium is Copyright (c) 2009-2010 by Appcelerator, Inc.
+ * and licensed under the Apache Public License (version 2)
+ */
+
+var Cloud = exports || {};
+var ACS = {};
+
+