+++ commonjs/src/sdk/namespace.js
@@ -1,0 +1,9 @@
+var com = {};
+com.cocoafish = {};
+com.cocoafish.constants = {};
+com.cocoafish.js = {};
+com.cocoafish.js.sdk = {};
+com.cocoafish.js.sdk.utils = {};
+
+com.cocoafish.sdk = {};
+com.cocoafish.sdk.url = {};